// src/utils/types.ts
export enum CurrencyEnum {
  USD = 'USD',
  JPY = 'JPY',
  CNY = 'CNY',
  CHF = 'CHF',
  CAD = 'CAD',
  MXN = 'MXN',
  INR = 'INR',
  BRL = 'BRL',
  RUB = 'RUB',
  KRW = 'KRW',
  IDR = 'IDR',
  TRY = 'TRY',
  SAR = 'SAR',
  SEK = 'SEK',
  NGN = 'NGN',
  PLN = 'PLN',
  ARS = 'ARS',
  NOK = 'NOK',
  TWD = 'TWD',
  IRR = 'IRR',
  AED = 'AED',
  COP = 'COP',
  THB = 'THB',
  ZAR = 'ZAR',
  DKK = 'DKK',
  MYR = 'MYR',
  SGD = 'SGD',
  ILS = 'ILS',
  HKD = 'HKD',
  EGP = 'EGP',
  PHP = 'PHP',
  CLP = 'CLP',
  PKR = 'PKR',
  IQD = 'IQD',
  DZD = 'DZD',
  KZT = 'KZT',
  QAR = 'QAR',
  CZK = 'CZK',
  PEN = 'PEN',
  RON = 'RON',
  VND = 'VND',
  BDT = 'BDT',
  HUF = 'HUF',
  UAH = 'UAH',
  AOA = 'AOA',
  MAD = 'MAD',
  OMR = 'OMR',
  CUC = 'CUC',
  BYR = 'BYR',
  AZN = 'AZN',
  LKR = 'LKR',
  SDG = 'SDG',
  SYP = 'SYP',
  MMK = 'MMK',
  DOP = 'DOP',
  UZS = 'UZS',
  KES = 'KES',
  GTQ = 'GTQ',
  URY = 'URY',
  HRV = 'HRV',
  MOP = 'MOP',
  ETB = 'ETB',
  CRC = 'CRC',
  TZS = 'TZS',
  TMT = 'TMT',
  TND = 'TND',
  PAB = 'PAB',
  LBP = 'LBP',
  RSD = 'RSD',
  LYD = 'LYD',
  GHS = 'GHS',
  YER = 'YER',
  BOB = 'BOB',
  BHD = 'BHD',
  CDF = 'CDF',
  PYG = 'PYG',
  UGX = 'UGX',
  SVC = 'SVC',
  TTD = 'TTD',
  AFN = 'AFN',
  NPR = 'NPR',
  HNL = 'HNL',
  BIH = 'BIH',
  BND = 'BND',
  ISK = 'ISK',
  KHR = 'KHR',
  GEL = 'GEL',
  MZN = 'MZN',
  BWP = 'BWP',
  PGK = 'PGK',
  JMD = 'JMD',
  XAF = 'XAF',
  NAD = 'NAD',
  ALL = 'ALL',
  SSP = 'SSP',
  MUR = 'MUR',
  MNT = 'MNT',
  NIO = 'NIO',
  LAK = 'LAK',
  MKD = 'MKD',
  AMD = 'AMD',
  MGA = 'MGA',
  XPF = 'XPF',
  TJS = 'TJS',
  HTG = 'HTG',
  BSD = 'BSD',
  MDL = 'MDL',
  RWF = 'RWF',
  KGS = 'KGS',
  GNF = 'GNF',
  SRD = 'SRD',
  SLL = 'SLL',
  XOF = 'XOF',
  MWK = 'MWK',
  FJD = 'FJD',
  ERN = 'ERN',
  SZL = 'SZL',
  GYD = 'GYD',
  BIF = 'BIF',
  KYD = 'KYD',
  MVR = 'MVR',
  LSL = 'LSL',
  LRD = 'LRD',
  CVE = 'CVE',
  DJF = 'DJF',
  SCR = 'SCR',
  SOS = 'SOS',
  GMD = 'GMD',
  KMF = 'KMF',
  STD = 'STD',
  XRP = 'XRP',
  AUD = 'AUD',
  BGN = 'BGN',
  BTC = 'BTC',
  JOD = 'JOD',
  GBP = 'GBP',
  ETH = 'ETH',
  EUR = 'EUR',
  LTC = 'LTC',
  NZD = 'NZD'
}